modify migration by changing attributes data type to unsigned int in order not to allow negative

+++ database/migrations/2020_01_31_101038_create_item_issued_table.php
@@ -21,7 +21,7 @@ class CreateItemIssuedTable extends Migration
             $table->uuid('item_id');
             $table->foreign('item_id')->references('id')->on('items')->onUpdate('cascade')->onDelete('cascade');
 
-            $table->float('quantity');
+            $table->unsignedInteger('quantity');
 
             $table->timestamps();
         });
